+++ tests/fixtures/configs/duplicate_outputs.yaml
