+++ tests/fixtures/configs/for.yaml
@@ -14,7 +14,7 @@ var:
 job:
   - for: v, range(1, var.counter)
     do:
-      - task: example
+      - task: for_example
         python: |
               excitement_level = ''
               for i in range(1, excited):
