+++ tests/integration/schema_test.py
@@ -1,15 +1,7 @@
 from unittest import TestCase
-import os
 from orca.core.config import OrcaConfig, OrcaConfigException  # noqa: F401
 
-fixture_path = os.path.dirname(__file__)
-
-
-def get_config(file_name):
-    path = os.path.join(fixture_path, 'fixtures', 'schema', file_name)
-    with open(path, "r") as p:
-        return OrcaConfig.create(p)
-
+from tests.util import get_config
 
 class OrcaSchemaTest(TestCase):
     pass
